refactor error modules
