+++ src/tests/filter_funcs_test.rs
@@ -1,5 +1,5 @@
 use crate::column::{Column, ColumnDataType};
-use crate::table::{errors::Error, NestedCondition, Table};
+use crate::table::{table_errors::Error, NestedCondition, Table};
 
 #[test]
 fn test_filter_with_nested_conditions() {
